Using session
I used the concept of session to get information accessible on multiple pages. In this case, the user's first name displayed all pages after the user has filled out the form. The name of user is displayed only if the sessions variable has been set. Finally, a forget me link was created to allows the user to end/destroy the session.